# Core/contracts/openapi/daemon-api.yaml
openapi: "3.1.0"
info:
  title: ScientificState Local Daemon API
  version: "0.1.0"
  description: |
    Local execution daemon API — localhost only.
    The daemon performs scientific computation; it does NOT assert scientific claims.
    All compute outputs require researcher review before claim evaluation.

    Constitutional boundaries:
    - P7: Non-delegation of scientific authority — daemon executes, human evaluates
    - P1: Raw data immutability — daemon never mutates stored data
    - Localhost boundary: No remote daemon access in M1
  contact:
    name: ScientificState
    url: https://scientificstate.org
  license:
    name: Constitutional Science Infrastructure License
servers:
  - url: "http://127.0.0.1:9473"
    description: Local daemon (Execution_Plan_Phase0.md §4.1 — port 9473)
...
paths:
...
  /modules/install:
    post:
      operationId: installModule
      tags: [modules]
      summary: Install a module
      description: |
        Installs a domain module after verifying its manifest signature.
        Constitutional constraint: unsigned modules are ALWAYS rejected (Phase1-B).
        Revoked modules are rejected before installation begins.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleInstallRequest"
      responses:
        "201":
          description: Module installed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleInstallResult"
        "400":
          description: Invalid request, bad manifest, or checksum mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signature verification failed or module revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Module already installed at this version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /modules/{domain_id}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get installed module details
      parameters:
        - name: domain_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Module details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleEntry"
        "404":
          description: Module not installed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeModule
      tags: [modules]
      summary: Remove an installed module
      description: |
        Removes the module code directory.
        Constitutional constraint (P9 — reversibility): SSV, claim, and run
        data are NEVER deleted by module removal. data_preserved is always true.
      parameters:
        - name: domain_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Module removed (data preserved)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleRemoveResult"
        "404":
          description: Module not installed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /modules/search:
    get:
      operationId: searchModules
      tags: [modules]
      summary: Search modules in registry
      description: Search the module registry by keyword or category.
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query string
        - name: category
          in: query
          schema:
            type: string
          description: Filter by taxonomy field
      responses:
        "200":
          description: Matching modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModuleEntry"
  /modules/update:
    post:
      operationId: updateModule
      tags: [modules]
      summary: Update an installed module to a newer version
      description: Downloads and verifies the latest version, then replaces the installed module.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain_id]
              properties:
                domain_id:
                  type: string
                target_version:
                  type: string
                  description: Specific version to update to — latest if omitted
      responses:
        "200":
          description: Module updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleInstallResult"
        "400":
          description: Invalid request or downgrade attempted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Module not installed or version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /modules/suggest:
    post:
      operationId: suggestModules
      tags: [modules]
      summary: Auto-detect compatible domain modules for a file
      description: |
        Analyzes a file path and returns matching domain modules based on
        format_signatures in module manifests. Does NOT install anything.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_path]
              properties:
                file_path:
                  type: string
                  description: Absolute local file path to analyze
      responses:
        "200":
          description: Suggested modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModuleEntry"
  /modules/package:
    post:
      operationId: packageModule
      tags: [modules]
      summary: Package and sign a local domain module
      description: |
        Creates a distributable tarball from a local domain module directory,
        computes checksums, and signs the manifest. Publish is a portal operation — NOT daemon.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_path]
              properties:
                source_path:
                  type: string
                  description: Local directory path containing the domain module
                private_key_path:
                  type: string
                  description: Path to Ed25519 private key for signing
      responses:
        "200":
          description: Module packaged and signed
          content:
            application/json:
              schema:
                type: object
                required: [tarball_path, tarball_hash, manifest_path]
                properties:
                  tarball_path:
                    type: string
                  tarball_hash:
                    type: string
                  manifest_path:
                    type: string
                  signature:
                    type: string
        "400":
          description: Invalid module directory or missing manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /modules/revoke:
    post:
      operationId: revokeModule
      tags: [modules]
      summary: Revoke a module version
      description: |
        Marks a module version as revoked. Revoked modules cannot be installed
        or activated. Existing installations are deactivated on next daemon restart.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain_id, version, reason]
              properties:
                domain_id:
                  type: string
                version:
                  type: string
                reason:
                  type: string
                  description: Human-readable revocation reason
      responses:
        "200":
          description: Module revoked
          content:
            application/json:
              schema:
                type: object
                required: [success, domain_id, version]
                properties:
                  success:
                    type: boolean
                  domain_id:
                    type: string
                  version:
                    type: string
        "404":
          description: Module or version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ModuleEntry:
      type: object
      required: [domain_id, version, install_path]
      properties:
        domain_id:
          type: string
        version:
          type: string
          pattern: '^\d+\.\d+\.\d+$'
        install_path:
          type: string
          description: Absolute path to the installed module directory
    ModuleInstallRequest:
      type: object
      required: [manifest_b64, package_b64, public_key_b64]
      properties:
        manifest_b64:
          type: string
          description: Base64-encoded manifest JSON bytes
        package_b64:
          type: string
          description: Base64-encoded module package bytes
        public_key_b64:
          type: string
          description: Base64-encoded DER Ed25519 public key
    ModuleInstallResult:
      type: object
      required: [success, domain_id, version]
      properties:
        success:
          type: boolean
        domain_id:
          type: string
        version:
          type: string
        install_path:
          type: string
    ModuleRemoveResult:
      type: object
      required: [success, domain_id, data_preserved]
      properties:
        success:
          type: boolean
        domain_id:
          type: string
        data_preserved:
          type: boolean
          description: Always true — SSV/claim data is never deleted (P9)
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        detail:
          type: object